ONCO-51: Fixed issue with the FHIR codeable concept resource
